backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
numpy==1.24.3
sympy==1.12
matplotlib==3.7.1
scipy==1.10.1
alembic==1.10.4
psycopg2-binary==2.9.6
firebase-admin==6.1.0